+++ UI/Frontend/js/managers/TransportPathManager.js
@@ -1,60 +1,13 @@
 import { State } from "../core/state.js";
+import { TransportPathDeductionInterface } from "../interfaces/TransportPathDeductionInterface.js";
 
 export class TransportPathManager {
     
     constructor(connectionRenderer) {
         this.renderer = connectionRenderer;
+        this.transportPathDeductionInterface= TransportPathDeductionInterface()
     }
 
-    getRoomLabel(roomId) {
-        return State.rooms.get(roomId).label;
-    }
-
-    path_input(){
-        const aperture_data = [];
-        State.apertures.forEach(ap => {   
-            const rooms = ap.rooms;   
-            if(ap.grounded){
-
-                const originRoom = this.getRoomLabel(rooms[0]);
-                const destination = rooms[1];
-                aperture_data.push({
-                origin: originRoom,
-                destination: destination,
-                id: ap.id
-                });
-
-            }
-            else if (rooms.length === 2) {
-                const originRoom = this.getRoomLabel(rooms[0]);
-                const destRoom = this.getRoomLabel(rooms[1]);
-                aperture_data.push({
-                origin: originRoom,
-                destination: destRoom,
-                id: ap.id
-                });
-            }
-        });
-        return aperture_data
-    }
-
-    
-    async invoke(apertures) {
-        const result = await fetch(`/transport/paths`, {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({apertures})
-        });
-        return result.json();
-    }
-
-    deducePaths(){
-        const input = this.path_input();
-        const pathPromise = this.invoke(input);
-        return pathPromise
-    }
 
     start() {
         if (State.transportPathMode) return;
@@ -62,7 +15,7 @@ export class TransportPathManager {
         
         State.ui.tranportPathList.style.display = "flex";
 
-        const pathPromise = this.deducePaths();
+        const pathPromise = this.transportPathDeductionInterface.deducePaths();
         pathPromise.then((paths) => {
             for (let i = 0; i < paths.length; i++) {
                 const tile = document.createElement("div");
